make_ext4fs very old and nasty bug fix.
Fixed a very, very, very old bug in make_ext4fs that makes it generate
invalid image files if one of the source files is found to be unreadable
during image creation. Now the tool will a) try to make it readable and
b) if it fails, the tool will display an error message and remove image.